refactor(scraper): implement banner script as action
